third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/security/contentSecurityPolicy/require-sri-for/require-sri-for-svg-script-blocked-expected.txt

CONSOLE ERROR: Refused to load the script 'http://127.0.0.1:8000/security/contentSecurityPolicy/require-sri-for/not-ran.js' because 'require-sri-for' directive requires integrity attribute be present for all scripts.
This is a testharness.js-based test.
PASS Script without integrity generates reports. 
PASS Script that requires integrity executes and does not generate a violation report. 
PASS Request to script without integrity is blocked, and generates violation report 
Harness: the test ran to completion.
